+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## 0.1.1
 
-* Add expression for (alias _name_) in (deps).
+* Add expression for `(alias _name_)` in `(deps)`.
+* Add `(with-stdin-from)` action.
 
 ## 0.1.0 (2022-10-14)